export router function in different folder not in config
