+++ index.js
@@ -1,4 +1,4 @@
-const router = require('./config/routes');
+const router = require('./routes/index');
 const fs = require('fs');
 
 const env = process.env.NODE_ENV || 'development';
@@ -8,6 +8,6 @@ const express = require('express');
 const app = express();
 
 require('./config/express')(app);
-require('./config/routes')(app);
+require('./routes/index')(app);
 
 app.listen(config.port, console.log(`Listening on port ${config.port}! Now its up to you...`));
